backend/pyproject.toml
[project]
name = "ai-calling-app-backend"
version = "0.1.0"
description = "AI Calling Application Backend with FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "supabase>=2.13.0",
    "python-multipart>=0.0.12",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "PyPDF2>=3.0.1",
    "pdfplumber>=0.11.4",
    "python-docx>=1.1.2",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
